requirements.txt
polygon-api-client
websocket-client
matplotlib
finplot
stockstats
pandas_datareader
mplfinance
PyPortfolioOpt
